# docs/openapi.yaml
openapi: 3.0.0
info:
  title: MyTrello API
  version: 0.1.9
  contact:
    name: Shagalov Veaceslav
    url: t.me/slava_shagalov
servers:
  - url: http://localhost/api/v1
    description: Main server
...
paths:
...
  /users/{id}:
    parameters:
      - schema:
          type: integer
        name: id
        in: path
        required: true
    get:
      summary: Returns user with specified id
      responses:
        "200":
          description: A JSON with user data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/ErrUnauthorized"
        "404":
          $ref: "#/components/responses/ErrUserNotFound"
        "500":
          $ref: "#/components/responses/ErrInternalServerError"
      tags:
        - Users
      security:
        - cookieAuth: [ ]
    patch:
      tags:
        - Users
      summary: Partial update of user
      description: Partial update of user
      requestBody:
        description: User data to update
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PartialUpdateUser"
      responses:
        "200":
          description: Updated user data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/ErrResponseBadRequest"
        "401":
          $ref: "#/components/responses/ErrUnauthorized"
        "403":
          $ref: "#/components/responses/ErrForbidden"
        "404":
          $ref: "#/components/responses/ErrUserNotFound"
      security:
        - cookieAuth: [ ]
  /users/{id}/avatar:
    parameters:
      - schema:
          type: integer
        name: id
        in: path
        required: true
    put:
      tags:
        - Users
      summary: Update of user avatar
      description: Update of user avatar
      requestBody:
        description: Multipart form with avatar file
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/UpdateUserAvatar"
      responses:
        "200":
          description: A JSON object of updated user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/ErrResponseBadRequest"
        "401":
          $ref: "#/components/responses/ErrUnauthorized"
        "403":
          $ref: "#/components/responses/ErrForbidden"
        "404":
          $ref: "#/components/responses/ErrUserNotFound"
      security:
        - cookieAuth: [ ]
...
  /boards/{id}/lists:
    parameters:
      - schema:
          type: integer
        name: id
        in: path
        required: true
    post:
      tags:
        - Boards
      summary: Create a new list
      description: Create a new list
      requestBody:
        description: A JSON object containing the data to create list.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateList"
      responses:
        "200":
          description: A JSON object of created list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/List"
        "400":
          $ref: "#/components/responses/ErrResponseBadRequest"
        "401":
          $ref: "#/components/responses/ErrUnauthorized"
        "500":
          $ref: "#/components/responses/ErrInternalServerError"
      security:
        - cookieAuth: [ ]
    get:
      tags:
        - Boards
      summary: Returns all lists with cards by board id
      description: Returns all lists with cards of current user
      responses:
        "200":
          description: A JSON with lists array
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListsWithCards"
        "401":
          $ref: "#/components/responses/ErrUnauthorized"
        "500":
          $ref: "#/components/responses/ErrInternalServerError"
      security:
        - cookieAuth: [ ]
  /lists:
    parameters:
      - name: title
        in: query
        description: Query string for list title
        required: false
        schema:
          type: string
    get:
      tags:
        - Lists
      summary: Returns user lists by filter parameters
      description: Returns user lists by filter parameters
      responses:
        "200":
          description: A JSON object with filtered user lists.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lists"
        "400":
          $ref: "#/components/responses/ErrResponseBadRequest"
        "401":
          $ref: "#/components/responses/ErrUnauthorized"
        "500":
          $ref: "#/components/responses/ErrInternalServerError"
      security:
        - cookieAuth: [ ]
  /lists/{id}:
    parameters:
      - schema:
          type: integer
        name: id
        in: path
        required: true
    get:
      tags:
        - Lists
      summary: Returns list by id
      description: Returns list by id
      responses:
        "200":
          description: A JSON object of list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/List"
        "400":
          $ref: "#/components/responses/ErrResponseBadRequest"
        "401":
          $ref: "#/components/responses/ErrUnauthorized"
        "404":
          $ref: "#/components/responses/ErrListNotFound"
        "500":
          $ref: "#/components/responses/ErrInternalServerError"
      security:
        - cookieAuth: [ ]
    patch:
      tags:
        - Lists
      summary: Partial update of list
      description: Partial update of list
      requestBody:
        description: A JSON object containing the data to update list.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PartialUpdateList"
      responses:
        "200":
          description: A JSON object of updated list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/List"
        "400":
          $ref: "#/components/responses/ErrResponseBadRequest"
        "401":
          $ref: "#/components/responses/ErrUnauthorized"
        "403":
          $ref: "#/components/responses/ErrForbidden"
        "404":
          $ref: "#/components/responses/ErrListNotFound"
      security:
        - cookieAuth: [ ]
    delete:
      tags:
        - Lists
      summary: Delete list
      description: Delete list
      responses:
        "204":
          description: List deleted successfully.
        "400":
          $ref: "#/components/responses/ErrResponseBadRequest"
        "401":
          $ref: "#/components/responses/ErrUnauthorized"
        "404":
          $ref: "#/components/responses/ErrListNotFound"
      security:
        - cookieAuth: [ ]
...
components:
  schemas:
    Id:
      type: integer
      required: true
      example: 3
    Email:
      type: string
      example: slava@vk.com
...
    Username:
      type: string
      minLength: 4
      maxLength: 40
      example: slava51
    Name:
      type: string
      minLength: 1
      maxLength: 60
      example: Slava
...
    MediaSource:
      type: string
      example: https://s3.trello.ru/uploaded_picture.jpg
    Time:
      type: string
      example: "11.07.2023 12:23:45"
...
    User:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/Id"
        username:
          $ref: "#/components/schemas/Username"
        email:
          $ref: "#/components/schemas/Email"
        name:
          $ref: "#/components/schemas/Name"
        avatar:
          $ref: "#/components/schemas/MediaSource"
        created_at:
          $ref: "#/components/schemas/Time"
        updated_at:
          $ref: "#/components/schemas/Time"
    PartialUpdateUser:
      type: object
      properties:
        username:
          type: string
          required: false
          example: "slava51"
        email:
          type: string
          required: false
          example: "slava@vk.com"
        name:
          type: string
          required: false
          example: "Slava"
    UpdateUserAvatar:
      type: object
      properties:
        avatar:
          type: string
          format: binary
...
    List:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/Id"
        board_id:
          $ref: "#/components/schemas/Id"
        title:
          type: string
          example: "API"
        position:
          type: integer
          example: 4
        created_at:
          $ref: "#/components/schemas/Time"
        updated_at:
          $ref: "#/components/schemas/Time"
    ListWithCards:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/Id"
        board_id:
          $ref: "#/components/schemas/Id"
        title:
          type: string
          example: "API"
        position:
          type: integer
          example: 4
        created_at:
          $ref: "#/components/schemas/Time"
        updated_at:
          $ref: "#/components/schemas/Time"
        cards:
          type: array
          items:
            $ref: "#/components/schemas/Card"
    Lists:
      type: object
      properties:
        lists:
          type: array
          items:
            $ref: "#/components/schemas/List"
    ListsWithCards:
      type: object
      properties:
        lists:
          type: array
          items:
            $ref: "#/components/schemas/ListWithCards"
    CreateList:
      type: object
      properties:
        title:
          type: string
          example: "API"
      example: {
        "title": "API",
      }
    PartialUpdateList:
      type: object
      properties:
        title:
          type: string
          required: false
          example: "API"
        position:
          type: integer
          required: false
          example: 4
        board_id:
          required: false
          type: integer
          example: 3
    Card:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/Id"
        list_id:
          $ref: "#/components/schemas/Id"
        title:
          type: string
          example: "REST"
        content:
          type: string
          example: "RESTFUL API DOCS: https://restfulapi.net/"
        position:
          type: integer
          example: 5
        created_at:
          $ref: "#/components/schemas/Time"
        updated_at:
          $ref: "#/components/schemas/Time"
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
  responses:
    ErrResponseBadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            description: Bad request
            properties:
              error:
                type: string
                example: "{{reason}}"
    ErrUnauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            description: Unauthorized
            properties:
              error:
                type: string
                example: "unauthorized"
    ErrUserNotFound:
      description: User not found
      content:
        application/json:
          schema:
            type: object
            description: User not found
            properties:
              error:
                type: string
                example: "user not found"
...
    ErrListNotFound:
      description: List not found
      content:
        application/json:
          schema:
            type: object
            description: List not found
            properties:
              error:
                type: string
                example: "list not found"
...
    ErrForbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            type: object
            description: Access denied
            properties:
              error:
                type: string
                example: "access denied"
    ErrInternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            type: object
            description: Internal Server Error
            properties:
              error:
                type: string
                example: "{{reason}}"